added graphics to mostly evereyting

+++ ZombieSmash/Assets/TimeElapsedDisp.cs
@@ -6,7 +6,7 @@ public class TimeElapsedDisp : MonoBehaviour {
 	public Text TimeText;
 	// Update is called once per frame
 	void Update () {
-		TimeText.text = (GM.Instance.timeTotal).ToString();
+		TimeText.text = ((int)GM.Instance.timeTotal).ToString();
 	}
 }
